Update server setup files
